controller와 dto 코드 정리
