--- src/main/java/com/example/gitboard/dto/UserProfileDto.java
@@ -1,12 +1,0 @@
-package com.example.gitboard.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class UserProfileDto {
-    private Long id;
-    private String username;
-    private String role;
-}
