+++ src/main/java/com/example/gitboard/service/BoardService.java
@@ -59,9 +59,9 @@ public class BoardService {
         boardRepository.delete(board);
     }
 
-    public List<BoardResponseDto> findByUser(User user) {
+    public List<BoardResponseDto> findMyBoards(User user) {
         return boardRepository.findAllByUser(user).stream()
                 .map(BoardMapper::toDto)
-                .collect(Collectors.toList());
+                .toList();
     }
 }
